requirements.txt
PyQt6==6.6.1
PyMuPDF==1.23.8
python-pptx==0.6.23
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pyaudio==0.2.11
keyboard==0.13.5 